Implementing the new `processTransaction` function
Implementing the `processTransaction` function, which handles the response from Gr4vy after the Tx req is done.
This function sets/updates all the data needed for the magento order with the latest data received from Gr4vy. If the Tx is not successful, it also creates a magento order with cancelled status and handles stopping the order confirmation going to the real customer. To achieve this, we are changing the `customerEmail` value to a fake email which is being used as a metadata for the `OrderEmailPlugin` to stop email going out. We then update the `customerEmail` to it's original value again.