Changing the order placement logic in the frontend
The new logic sends a request to the `TransactionRepositoryInterface` to process Gr4vy response. If the tx is successful, we set all the data required for the Magento Order in the backend and then we place an order by calling the native Magento JS function that handles order placement.
If the Tx is not successful, the BE handles creating a Magento order with `canceled` status in the backend.