Fixing multiple tab issue
I'm cloning the basket after embed is initialised and storing the new basket id in the session.
Once the Gr4vy Tx is done and before placing a Magento order, I am updating/overriding the user basket with what was cloned before the Tx to make sure what the user paid for and what we have in Gr4vy matches what is in the Magento order.